fix: offline services restructure + INTEGER overflow + sync column mismatch
- Restructured offline services into core/, sync/, cache/, documents/
- Fixed all broken import paths after folder restructure
- Fixed INTEGER → BIGINT in document number SQL (overflow for 12-digit numbers)
- Fixed sync.py p.category → p.category_id to match database schema
- Added offlineStorage.js stub for legacy components
- Improved batch not found soft validation for offline invoices
- Bumped service worker cache version to v3

+++ backend/app/api/routes/documents.py
@@ -43,7 +43,7 @@ async def generate_document_number(
         
         # Get last number for this document type
         result = db.execute(text("""
-            SELECT MAX(CAST(REGEXP_REPLACE(document_number, '[^0-9]', '', 'g') AS INTEGER)) as last_number
+            SELECT MAX(CAST(REGEXP_REPLACE(document_number, '[^0-9]', '', 'g') AS BIGINT)) as last_number
             FROM (
                 SELECT invoice_number as document_number FROM sales.invoices WHERE org_id = :org_id
                 UNION ALL
